Lighten Megatron orchestration imports
